+++ CHANGELOG.md
@@ -1,4 +1,5 @@
 ## v0.2.0
+* Toggle response builder
 * Expand body's file path
 * Implement File Body response
 * UI bug fixes & refactor